+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 ===================
 
+* 1.0.1 (YYYY-MM-DD)
+ * Exceptions are now from the global and correct exception type
+ * The complete result from the API call will be returned instead of only the rate or amount
+ * PHPUnit tests
+
 * 1.0.0 (2014-12-12)
  * First stable version for production
  * Supported API endpoints: `get`, `convert`, `historical`, `locale`